requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
face-recognition==1.3.0
dlib-bin==19.24.1
SpeechRecognition==3.10.0
pyaudio==0.2.11
python-dateutil==2.8.2
pillow==10.0.1
pyttsx3==2.90
# Google Cloud dependencies
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0